Fix bugs

+++ cmd/site/main.go
@@ -48,7 +48,7 @@ func main() {
 			http.Error(w, err.Error(), http.StatusBadRequest)
 			return
 		}
-		recv := strconv.FormatUint(ntpal.GetSystemTime(), 10)
+		recv := strconv.FormatUint(uint64(ntpal.GetSystemTime()), 10)
 
 		syncResponse := SyncResponse{
 			Orig: syncRequest.Orig,
@@ -58,7 +58,7 @@ func main() {
 
 		encoder := json.NewEncoder(w)
 
-		syncResponse.Xmt = strconv.FormatUint(ntpal.GetSystemTime(), 10)
+		syncResponse.Xmt = strconv.FormatUint(uint64(ntpal.GetSystemTime()), 10)
 		encoder.Encode(syncResponse)
 	})
